+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix: Linked databases are handled as `ChildDatabase` objects and no longer resolved, issue #110.
 - Chg: Have a `User` hierarchy with `Bot`, `Person`, etc. to reflect the actual Notion API.
 - Doc: Completed missing parts in the documentation and example for file uploads.
+- New: Allow uploading and appending a file as block to a page.
 
 ## Version 0.9, 2025-09-01